Moved PTransforms into bigtable package in templates
