+++ v2/googlecloud-to-googlecloud/src/main/java/com/google/cloud/teleport/v2/templates/bigtablechangestreamstogcs/BigtableChangeStreamMutationToChangelogEntryFn.java
@@ -1,12 +1,16 @@
-package com.google.cloud.teleport.v2.transforms;
+package com.google.cloud.teleport.v2.templates.bigtablechangestreamstogcs;
 
 import com.google.cloud.bigtable.data.v2.models.ChangeStreamMutation;
-import com.google.cloud.teleport.v2.templates.bigtablechangestreamstogcs.BigtableUtils;
 import java.nio.charset.Charset;
 import java.util.HashSet;
 import java.util.List;
 import org.apache.beam.sdk.transforms.SimpleFunction;
 
+/**
+ * Class used as a {@link org.apache.beam.sdk.transforms.PTransform} to get
+ * valid {@link com.google.cloud.teleport.bigtable.ChangelogEntry} objects based on the parameters
+ * provided by the pipeline.
+ */
 public class BigtableChangeStreamMutationToChangelogEntryFn extends
     SimpleFunction<ChangeStreamMutation, List<com.google.cloud.teleport.bigtable.ChangelogEntry>> {
